Extract functions to modules
