+++ lib/utils/password.js
@@ -1,0 +1,22 @@
+const rl = require('readline');
+const logger = require('../logger');
+
+const readPassword = () => {
+  return new Promise((resolve, reject) => {
+    const reader = rl.createInterface({ input: process.stdin, output: process.stdout });
+    reader.stdoutMuted = true;
+    reader.question('Password: ', password => {
+      reader.close();
+      logger.Info('');
+      resolve(password);
+    });
+
+    reader._writeToOutput = stringToWrite => {
+      (reader.stdoutMuted && reader.output.write('*')) || reader.output.write(stringToWrite);
+    };
+  });
+};
+
+module.exports = {
+  readPassword: readPassword
+};
